Move messaging transformers to an ExtensionTransformerRegistration. Migrate to chained transformers in preparation for new model version.
